agropest12_project/requirements.txt

torch>=2.2.0
torchvision>=0.17.0
tqdm>=4.66.0
numpy>=1.26.0
opencv-python>=4.9.0.80
Pillow>=10.0.0
pyyaml>=6.0.1
matplotlib>=3.8.0
scikit-learn>=1.3.0
